Add basic classes for colors.

+++ Lab1/rgbcolor.cpp
@@ -1,12 +1,12 @@
 #include "rgbcolor.h"
 
-RGBColor::RGBColor() : r_(0), g_(0), b_(0)
+RGBColor::RGBColor ()
+    : r_ (0), g_ (0), b_ (0)
 {
-
 }
 
-RGBColor::RGBColor(int r, int g, int b) : r_(r), g_(g), b_(b)
+RGBColor::RGBColor (float r, float g, float b)
+    : r_ (r), g_ (g), b_ (b)
 {
-
 }
